+++ src/plone/app/multilingual/browser/templates/dexterity_edit.pt
@@ -20,11 +20,11 @@
            value context/absolute_url;
          "
   />
-  <input id="gtranslate_service_available"
+  <input id="translation_service_available"
          type="hidden"
          value=""
          tal:attributes="
-           value pamutils/gtenabled;
+           value pamutils/translations_enabled;
          "
   />
