+++ src/plone/app/multilingual/profiles/default/metadata.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <metadata>
-  <version>1002</version>
+  <version>1003</version>
   <dependencies>
     <dependency>profile-plone.app.dexterity:default</dependency>
   </dependencies>
