+++ src/plone/app/multilingual/profiles/default/registry.xml
@@ -32,7 +32,6 @@
     <value key="redirect_babel_view">True</value>
     <value key="bypass_languageindependent_field_permission_check">False</value>
     <value key="buttons_babel_view_up_to_nr_translations">7</value>
-    <value key="google_translation_key" />
     <value key="selector_lookup_translations_policy">closest</value>
   </records>
